[android] Refactor MinRequiredFramesTester
b/327287075
